Working with maps, simple visualization, random algorithm
