+++ src/main/kotlin/com/aditshah/distributed/CoordinateReporting.kt
@@ -1,0 +1,5 @@
+package com.aditshah.distributed
+
+fun main() {
+
+}
